refactor: further separate API logic from CLI, and create submodules for each subcommand
